Bug #1485634 add UT-IT cases
[root cause  ] add UT-IT cases
[changes     ] Camera UT-IT framework
[side effects]no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:yes
[board]:ums512
[test case]:camera
[reviewers   ] hongcheng.li
[change_type ] feature
[tag_product ] common

+++ test/framework/json/src/IParseJson.cpp
@@ -20,7 +20,7 @@ void IParseJson::PrintValueTree(Json::Value &value, IParseJson *pParent,
         pobj->DealJsonNode(strkey, value.asUInt());
         break;
     case Json::realValue:
-        pobj->DealJsonNode(strkey, value.asDouble());
+        pobj->DealJsonNode(strkey, value.asFloat());
         break;
     case Json::stringValue:
         pobj->DealJsonNode(strkey, value.asString());
